Move existing WMI code to a separate file
