+++ src/endless/EndlessUsbTool.vcxproj.filters
@@ -141,6 +141,9 @@
     <ClInclude Include="ImageSigningKey.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="WMI.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="EndlessUsbTool.cpp">
@@ -237,6 +240,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="GeneralCode.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="WMI.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
