+++ src/endless/EndlessUsbToolDlg.h
@@ -433,8 +433,6 @@ private:
 
 	static CStringW GetSystemDrive();
 	static BOOL SetEndlessRegistryKey(HKEY parentKey, const CString &keyPath, const CString &keyName, CComVariant keyValue, bool createBackup = true);
-	static BOOL IsBitlockedDrive(const CString &drive);
-	static BOOL GetMachineInfo(CString &manufacturer, CString &model);
 
 	static CStringW GetExePath();
 	static BOOL AddUninstallRegistryKeys(const CStringW &uninstallExePath, const CStringW &installPath);
